refactor: action drawer
